base: the operating-system host behind the read-only seam, the volume's timestamps as 64-bit nanoseconds, the destroy gate attributing preempted slices, and a drift-aware ratchet
Phase 1 task 10 completes. crates/base (slates-base) implements the host seam over the
operating system: descriptor-relative rustix calls on Unix (openat with O_NOFOLLOW, fstat,
getdents with one statat per entry, pread, readlinkat, fstatfs for the timestamp-granularity
table), inotify on Linux and EVFILT_VNODE on macOS behind the same seam, and a path-relative
standard-library form on Windows without a watcher (fingerprints alone; recorded). The
structural test's write denylist gains rustix's write calls and open flags, so the crate's lint
wall is enforced, not promised. Tests run over the workspace's own crates/ directory (listings,
opens that never cross kinds, reads equal to the standard library's, an overlay costing one open)
and, in the Linux lane, over tmpfs (a held descriptor outliving a replacement, O_NOFOLLOW,
hints). Baseline (Apple M5 Max, APFS, 2026-09-05): listing 3.4 us per entry over 44,602
entries, a directory fingerprint 218 ns, an open and fstat 7.2 us, pread 260 ns, a small-class
copy-up 30 us, a large-class window 6.8 ms (the witness hashes the 5.6 MB file whole).

Timestamps are 64-bit nanoseconds everywhere (good to 2262, saturating): the inode is 264 bytes
and heap per file at 10^6 files is 418 bytes, below the 468 recorded before the deriver's home
was added. The volume bench's destroy gate excludes slices during which the process's
involuntary context-switch count moved (getrusage), and reports the per-unit cost from the
slices' own time and from one unbounded slice (14 ns). The deriver walks the head subtrees of
created and renamed directories (a directory created, filled and then renamed had lost its
entries; 1,800 generated histories since). A write into a window the volume already pinned no
longer runs the drift check. Slabs free large slots in place (discard) instead of moving a 4 KiB
block out. The ratchet tightens only past a row's own between-run drift and the baseline was
reset under that rule (60 gated rows, 0 regressions). The five bench examples are named per
crate; cargo warned their outputs collide.

+++ crates/vfs/src/inode.rs
@@ -34,13 +34,13 @@ pub struct Attrs {
   /// Size in bytes (the file length; a directory reports 0).
   pub size: u64,
   /// Access time, ns since the Unix epoch.
-  pub atime: i128,
+  pub atime: i64,
   /// Modification time.
-  pub mtime: i128,
+  pub mtime: i64,
   /// Change time.
-  pub ctime: i128,
+  pub ctime: i64,
   /// Birth time.
-  pub btime: i128,
+  pub btime: i64,
 }
 
 /// The base-plane fields of a base-backed body (§4.5 `Body::Base`), filled in by `slates-base`.
@@ -81,9 +81,9 @@ pub struct Fingerprint {
   /// Size.
   pub size: u64,
   /// Modification time, ns.
-  pub mtime_ns: i128,
+  pub mtime_ns: i64,
   /// Change time, ns.
-  pub ctime_ns: i128,
+  pub ctime_ns: i64,
   /// Mode.
   pub mode: u32,
 }
